pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cd-agency"
version = "0.6.0"
description = "AI-powered content design agents for UX writers, conversation designers, and product teams"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
keywords = [
    "content design",
    "UX writing",
    "AI agents",
    "microcopy",
    "conversation design",
    "content strategy",
    "accessibility",
    "WCAG",
    "error messages",
    "onboarding",
    "design system",
    "CLI",
    "MCP",
    "Figma",
    "FastAPI",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Other Audience",
    "Topic :: Software Development :: User Interfaces",
    "Topic :: Text Processing :: Linguistic",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "anthropic>=0.39.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "click>=8.1",
    "textual>=0.80.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.5",
]
api = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.25.0",
    "python-multipart>=0.0.6",
]
web = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.25.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
]
multi-model = [
    "openai>=1.3.0",
]
langgraph = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
]
vectors = [
    "chromadb>=0.4.18",
    "sentence-transformers>=2.2.2",
]
cloud = [
    "sqlalchemy[postgresql]>=2.0.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.25.0",
    "python-multipart>=0.0.6",
    "psycopg2-binary>=2.9.0",
]

[project.scripts]
cd-agency = "runtime.cli:main"

[project.urls]
Homepage = "https://github.com/adedayoagarau/cd-agency"
Repository = "https://github.com/adedayoagarau/cd-agency"
Issues = "https://github.com/adedayoagarau/cd-agency/issues"

[tool.setuptools.packages.find]
include = ["runtime*", "runtime.tui*", "tools*", "mcp*", "api*", "packages*"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.yaml", "*.yml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
